Some initial execution events.
I think the exact format of Events is going to be in flux for a little while now. I want it to have
some compatibility with AWS's GetExecutionHistory format, but possible enriched, but maybe not TOO
nosiy/redundant? For now, there's just a few event types, and we'll see where it goes.