Add reference paths, getting ready for ResultsPath support.
Reference Paths are a subset of JSONPath. While I was already using jsonpath_lib for all JSONPath
things, it doesn't seem to have any way to do the thing I need for ResultsPath: take a base Value, a
Path, and a new Value, and insert the new Value at the given path into the base.

That lack sent me on a whole adventure, where now I am pulling in the nom crate are parsing a subset
of JSONPath myself.

This commit just adds the necessary reference path code; in the next commit we will apply it to
ResultsSelector.