Rewrite of Task's input/output mechansim using a new StateIo trait.
This trait is setup to allow code reuse between all the different states, including those other than
Task/Parallel/Map that only allow a subset of the input/output modification fields. In those cases,
there's still a hidden value of, say, ResultPath, it just is a fixed constant.

This system also introduces additional type safety for paths. Previously, in order to preserve the
idea that not providing a path should correspond to None in an Option for ease of deserialization,
there were three cases, None, Some(Value::Null), and Some(Value::String). But the elimination of
other Value enum members was pushed to runtime. Now it's brought forward, making it an error at
deserialization time to have, say, an array for a path.