Implement StateIo for the State struct, lifting it out of the giant match.
This is one of the two places for code reuse in the reasoning for making StateIo in the first
place. This way, each state has uniform input handling, all in one place.

Maybe I'll figure out how to get `enum_dispatch` to work to reduce the giant boilerplate impl
StateIO for State, but for now it'll do.