Add subcrate that generated some Choice State boiler plate
Going to remove this in the next commit, but it was a nice learning exercise in proc
macros.

Adds a subcrate for a proc macro that generates the Choice state boilerplate around all the
variants of the form (Type)(Comparsion)(Path)?. Ultimately, this macro ended up being much bigger
and more complex than I was hoping for, comparable to the lines of code it was supposed to save. I
was hoping that it would be better than copy-pasting because it would centralize/DRY the business
logic, but, having written it, I think actually that having to maintain/debug the macro is worse
than having small repeated code blocks.